# Conflicts:
#	WalletOK/src/main/java/com/theost/walletok/presentation/wallet_details/transaction/TransactionActivity.kt
#	WalletOK/src/main/java/com/theost/walletok/presentation/wallet_details/transaction/TransactionCategoryFragment.kt
#	WalletOK/src/main/java/com/theost/walletok/presentation/wallet_details/transaction/TransactionEditFragment.kt
#	WalletOK/src/main/res/values/colors.xml

+++ WalletOK/src/main/java/com/theost/walletok/data/models/TransactionCreationModel.kt
@@ -7,7 +7,7 @@ import java.util.*
 @Parcelize
 class TransactionCreationModel(
     var id: Int? = null,
-    var value: Int? = null,
+    var value: Long? = null,
     var type: String? = null,
     var category: Int? = null,
     var currency: Currency? = null,
